requirements.txt
ollama
chromadb
pyyaml
rich
